+++ spade/behaviour.py
@@ -258,7 +258,7 @@ class CyclicBehaviour(object, metaclass=ABCMeta):
             msg.sender = str(self.agent.jid)
             logger.debug(f"Adding agent's jid as sender to message: {msg}")
         aioxmpp_msg = msg.prepare()
-        await self.agent.stream.send(aioxmpp_msg)
+        await self.agent.client.send(aioxmpp_msg)
         msg.sent = True
         self.agent.traces.append(msg, category=str(self))
 
@@ -474,7 +474,22 @@ class FSMBehaviour(CyclicBehaviour):
         behaviour.set_agent(self.agent)
         behaviour.receive = self.receive
         logger.info(f"FSM running state {self.current_state}")
-        await behaviour._start()
+        try:
+            await behaviour.on_start()
+        except Exception as e:
+            logger.error("Exception running on_start in state {}: {}".format(self, e))
+            self.kill(exit_code=e)
+        try:
+            await behaviour.run()
+        except Exception as e:
+            logger.error("Exception running state {}: {}".format(self, e))
+            self.kill(exit_code=e)
+        try:
+            await behaviour.on_end()
+        except Exception as e:
+            logger.error("Exception running on_start in state {}: {}".format(self, e))
+            self.kill(exit_code=e)
+
         dest = behaviour.next_state
 
         if dest: